mdk-devel-backup from 2016-03-31 02:32

+++ mdk/m23helper/signBaseSys.sh
@@ -2,7 +2,10 @@
 
 ls *.tar.7z | while read basesys
 do
-	cat "$basesys" | gpg --default-key 0x3305E63A --no-options --detach-sign --armor --textmode > "$basesys.sig"
+	if [ ! -e "$basesys.sig" ]
+	then
+		cat "$basesys" | gpg --default-key 0x3305E63A --no-options --detach-sign --armor --textmode > "$basesys.sig"
+	fi
 done
 
 rsync -Pya *.sig root@tux05:/var/www/basesys
